# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "skywatch station",
    "description": "The complete contract for a skywatch station: a receive-only SDR aviation\nmonitor that records VHF airband transmissions, transcribes them locally,\nflags the interesting ones, and attaches probable aircraft.\n\nReading the data:\n\n- `/status` \u2014 station health; `/frequencies` \u2014 the listening plan.\n- `/recordings` \u2014 the clip library; `/clips/interesting` \u2014 the good bits;\n  `/digest` \u2014 one day summarised for a \"today\" view.\n- `/runbook` and `/glossary` \u2014 station documents as Markdown.\n\nLive updates come from `WS /stream`. Every message is an envelope\n`{\"type\": <event>, \"payload\": <object>}` with three event types:\n`recording.new` (payload: a /recordings list item), `recording.updated`\n(same shape, sent when a clip's pipeline stage or verdicts change), and\n`status.changed` (payload: the /status shape, sent when station settings\nchange). The stream sends events only \u2014 anything a client sends is ignored.\n\nErrors are RFC 7807 problem details (`application/problem+json`) with a\nstable `code` field to switch on.\n\nThe API is unauthenticated by design: it is meant for a private LAN or\nTailscale network only, and must not be exposed to the public internet.\nTransmissions are received for personal monitoring; audio must not be\nrebroadcast.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/recordings": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Browse the clip library",
        "description": "Recorded transmissions, newest first, as summary cards: frequency, timing, latest transcript snippet, latest verdict, most plausible aircraft candidate, and feedback tallies. All filters combine. Date filters are inclusive and use the station's local day boundaries. `interesting` and `category` match each clip's latest verdict, so a reclassified clip follows its newest classification. `has_match` selects clips with (or without) aircraft candidates.",
        "operationId": "list_recordings_recordings_get",
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Earliest local day to include.",
              "title": "From Date"
            },
            "description": "Earliest local day to include."
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Latest local day to include.",
              "title": "To Date"
            },
            "description": "Latest local day to include."
          },
          {
            "name": "freq_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only clips from this frequency.",
              "title": "Freq Id"
            },
            "description": "Only clips from this frequency."
          },
          {
            "name": "interesting",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true for clips whose latest verdict is interesting; false for routine.",
              "title": "Interesting"
            },
            "description": "true for clips whose latest verdict is interesting; false for routine."
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ClassificationCategory"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only clips whose latest verdict has this category.",
              "title": "Category"
            },
            "description": "Only clips whose latest verdict has this category."
          },
          {
            "name": "has_match",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Whether clips must have aircraft candidates.",
              "title": "Has Match"
            },
            "description": "Whether clips must have aircraft candidates."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Clips per page.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Clips per page."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Clips to skip from the newest.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Clips to skip from the newest."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "One clip with full provenance",
        "description": "Everything the station knows about one clip: the latest transcript with engine provenance and confidence, the complete classification history (newest first \u2014 verdicts are appended, never overwritten), every ranked aircraft candidate, individual feedback entries, and any pipeline error the clip hit.",
        "operationId": "get_recording_recordings__recording_id__get",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingDetail"
                }
              }
            }
          },
          "404": {
            "description": "Unknown recording.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/audio": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Stream a clip's audio",
        "description": "The clip as MP3 (8 kHz mono as captured). Supports HTTP Range requests, so seeking in an audio player works. Returns 410 once retention has pruned the audio of an old routine clip \u2014 the clip's metadata remains available.",
        "operationId": "get_audio_recordings__recording_id__audio_get",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "audio/mpeg": {}
            }
          },
          "206": {
            "description": "Partial content for a Range request."
          },
          "404": {
            "description": "Unknown recording or missing file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Audio pruned by retention.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/reclassify": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Queue a clip for reclassification",
        "description": "Sends the clip back through the classifier. The pipeline worker picks it up on its next pass and appends a fresh verdict to the clip's history (nothing is overwritten). Returns 202 immediately; watch the stream or re-fetch the clip for the new verdict. Clips that have not been transcribed yet cannot be reclassified.",
        "operationId": "reclassify_recording_recordings__recording_id__reclassify_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReclassifyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown recording.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Clip has not been transcribed yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}/feedback": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Record a thumbs-up or thumbs-down",
        "description": "Stores a listener verdict on whether the clip was worth hearing. Verdicts accumulate \u2014 they are individual votes, not a toggle \u2014 and feed the all-time greatest-hits list in the digest.",
        "operationId": "create_feedback_recordings__recording_id__feedback_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recording Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResource"
                }
              }
            }
          },
          "404": {
            "description": "Unknown recording.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationCategory": {
        "type": "string",
        "enum": [
          "routine",
          "emergency",
          "urgency",
          "go_around",
          "medical",
          "fuel",
          "guard_activity",
          "unusual",
          "other"
        ],
        "title": "ClassificationCategory",
        "description": "Why a clip is (or is not) interesting."
      },
...
      "FeedbackCreate": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/FeedbackVerdict"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "verdict"
        ],
        "title": "FeedbackCreate",
        "description": "Request body for recording a listener verdict."
      },
      "FeedbackResource": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "recording_id": {
            "type": "integer",
            "title": "Recording Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/FeedbackVerdict"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "verdict",
          "note",
          "created_at"
        ],
        "title": "FeedbackResource",
        "description": "One listener verdict on a clip."
      },
      "FeedbackVerdict": {
        "type": "string",
        "enum": [
          "up",
          "down"
        ],
        "title": "FeedbackVerdict",
        "description": "Listener thumbs-up / thumbs-down on a clip."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Link": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          }
        },
        "type": "object",
        "required": [
          "href"
        ],
        "title": "Link",
        "description": "One HAL link: a relative URL on this API."
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "$ref": "#/components/schemas/APIErrorCode"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem details; the body of every error response."
      },
      "ReclassifyResponse": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "recording_id": {
            "type": "integer",
            "title": "Recording Id"
          },
          "stage": {
            "$ref": "#/components/schemas/RecordingStage"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "stage",
          "detail"
        ],
        "title": "ReclassifyResponse",
        "description": "Acknowledgement that a clip was queued for reclassification."
      },
      "RecordingDetail": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyRef"
          },
          "started_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Started At Utc"
          },
          "ended_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Ended At Utc"
          },
          "duration_s": {
            "type": "number",
            "title": "Duration S"
          },
          "stage": {
            "$ref": "#/components/schemas/RecordingStage"
          },
          "audio_available": {
            "type": "boolean",
            "title": "Audio Available",
            "description": "False once retention has pruned the audio; the row itself is kept."
          },
          "transcript_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Snippet",
            "description": "Opening of the latest transcript, or null while transcription is pending."
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassificationResource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Latest verdict, or null while classification is pending."
          },
          "top_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AircraftMatchResource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Most plausible aircraft candidate, or null when none was found."
          },
          "feedback": {
            "$ref": "#/components/schemas/FeedbackCounts"
          },
          "sample_rate": {
            "type": "integer",
            "title": "Sample Rate"
          },
          "file_path": {
            "type": "string",
            "title": "File Path",
            "description": "Audio location relative to the station's data root."
          },
          "stage_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Error",
            "description": "Most recent pipeline failure for this clip, if any stage struggled."
          },
          "audio_deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Deleted At"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptResource"
              },
              {
                "type": "null"
              }
            ]
          },
          "classifications": {
            "items": {
              "$ref": "#/components/schemas/ClassificationResource"
            },
            "type": "array",
            "title": "Classifications",
            "description": "Full verdict history, newest first."
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/AircraftMatchResource"
            },
            "type": "array",
            "title": "Matches",
            "description": "All candidates, best first."
          },
          "feedback_entries": {
            "items": {
              "$ref": "#/components/schemas/FeedbackResource"
            },
            "type": "array",
            "title": "Feedback Entries"
          }
        },
        "type": "object",
        "required": [
          "id",
          "frequency",
          "started_at_utc",
          "ended_at_utc",
          "duration_s",
          "stage",
          "audio_available",
          "transcript_snippet",
          "classification",
          "top_match",
          "feedback",
          "sample_rate",
          "file_path",
          "stage_error",
          "audio_deleted_at",
          "transcript",
          "classifications",
          "matches",
          "feedback_entries"
        ],
        "title": "RecordingDetail",
        "description": "A clip with its complete provenance."
      },
      "RecordingListResponse": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links",
            "description": "HAL links: self plus related resources and the actions currently available on this resource."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordingSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Matching clips across all pages."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RecordingListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}